updating home with routes
